Do not homogenise Line
It turns out that A + B + C could equal 0 in the line. This causes a
division by zero error when trying to get A + B + C = 1.

I have decided that there is no point in homogenising the line, thus two
lines with different A, B, C could possibly be the same. This is not
really a problem since the code never compares two lines.